# lxdweb/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "基于 Session 的认证",
        "title": "LXD Web 管理平台",
        "termsOfService": "https://github.com/xkatld/zjmf-lxd-server",
        "contact": {
            "name": "xkatld",
            "url": "https://github.com/xkatld/zjmf-lxd-server",
            "email": "xkatld@example.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://github.com/xkatld/zjmf-lxd-server/blob/main/LICENSE"
        },
        "version": "1.0.1"
    },
    "host": "localhost:3000",
    "basePath": "/",
    "paths": {
...
        "/api/nat": {
            "get": {
                "description": "查询所有NAT端口转发规则，支持按节点过滤",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "NAT管理"
                ],
                "summary": "获取NAT规则列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "节点ID",
                        "name": "node_id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功返回NAT规则列表",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "查询失败",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "post": {
                "description": "创建新的NAT端口转发规则",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "NAT管理"
                ],
                "summary": "创建NAT规则",
                "parameters": [
                    {
                        "description": "NAT规则参数",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.CreateNATRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "创建成功",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "参数错误",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "节点不存在",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "409": {
                        "description": "端口已被占用",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
        "/api/nat/sync": {
            "post": {
                "description": "同步指定节点的NAT规则信息",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "NAT管理"
                ],
                "summary": "同步NAT规则",
                "parameters": [
                    {
                        "type": "string",
                        "description": "节点ID",
                        "name": "node_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "同步任务已启动",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "缺少参数",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "节点不存在",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/nat/{id}": {
            "get": {
                "description": "根据ID获取NAT规则详情",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "NAT管理"
                ],
                "summary": "获取单个NAT规则",
                "parameters": [
                    {
                        "type": "string",
                        "description": "规则ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功返回NAT规则",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "NAT规则不存在",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "put": {
                "description": "更新NAT规则的描述或状态",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "NAT管理"
                ],
                "summary": "更新NAT规则",
                "parameters": [
                    {
                        "type": "string",
                        "description": "规则ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "更新参数",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.UpdateNATRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "更新成功",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "参数错误",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "NAT规则不存在",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "delete": {
                "description": "删除指定的NAT端口转发规则并清理数据库",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "NAT管理"
                ],
                "summary": "删除NAT规则",
                "parameters": [
                    {
                        "type": "string",
                        "description": "规则ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "删除成功",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "NAT规则或节点不存在",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.CreateNATRequest": {
            "type": "object",
            "required": [
                "container_hostname",
                "external_port",
                "internal_port",
                "node_id",
                "protocol"
            ],
            "properties": {
                "container_hostname": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "external_port": {
                    "type": "integer"
                },
                "internal_port": {
                    "type": "integer"
                },
                "node_id": {
                    "type": "integer"
                },
                "protocol": {
                    "type": "string",
                    "enum": [
                        "tcp",
                        "udp"
                    ]
                }
            }
        },
...
        "models.UpdateNATRequest": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "external_port": {
                    "type": "integer"
                },
                "internal_port": {
                    "type": "integer"
                },
                "protocol": {
                    "type": "string",
                    "enum": [
                        "tcp",
                        "udp"
                    ]
                },
                "status": {
                    "type": "string",
                    "enum": [
                        "active",
                        "inactive"
                    ]
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BasicAuth": {
            "type": "basic"
        }
    }
}